+++ src/engines/AbstractEngineChannel.cpp
@@ -890,8 +890,8 @@ namespace LinuxSampler {
      * @param delay - amount of microseconds in future (from now) when event shall be processed
      * @returns unique event ID of scheduled new event, or NULL on error
      */
-    event_id_t AbstractEngineChannel::ScheduleEventMicroSec(const Event* pEvent, int delay) {
-        dmsg(3,("AbstractEngineChannel::ScheduleEventMicroSec(Event.Type=%d,delay=%d)\n", pEvent->Type, delay));
+    event_id_t AbstractEngineChannel::ScheduleEventMicroSec(const Event* pEvent, int64_t delay) {
+        dmsg(3,("AbstractEngineChannel::ScheduleEventMicroSec(Event.Type=%d,delay=%lld)\n", pEvent->Type, delay));
         RTList<Event>::Iterator itEvent = pEvents->allocAppend();
         if (!itEvent) {
             dmsg(1,("AbstractEngineChannel::ScheduleEventMicroSec(): Event pool emtpy!\n"));
